chore: format 관련 util 메서드 utils/date/format 으로 이동
